# app/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "System Rejestracji Pacjenta Stomatologicznego API",
      "description": "API dla systemu rejestracji pacjenta stomatologicznego",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "/api",
        "description": "Development server"
      }
    ],
    "components": {
      "securitySchemes": {
        "jwt": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
    "paths": {
...
      "/integracja/znanylekarz/sync": {
        "post": {
          "tags": ["Integracje"],
          "summary": "Synchronizacja terminów z ZnanyLekarz",
          "security": [
            { "jwt": [] }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "lekarz_id": {"type": "integer"},
                    "data_od": {"type": "string", "format": "date"},
                    "data_do": {"type": "string", "format": "date"}
                  },
                  "required": ["lekarz_id"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Synchronizacja z ZnanyLekarz zakończona pomyślnie"
            },
            "400": {
              "description": "Błędne dane"
            }
          }
        }
      },
      "/integracja/nfz/ewus": {
        "post": {
          "tags": ["Integracje"],
          "summary": "Weryfikacja ubezpieczenia pacjenta w systemie eWUŚ",
          "security": [
            { "jwt": [] }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pacjent_id": {"type": "integer"},
                    "pesel": {"type": "string"}
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Weryfikacja eWUŚ zakończona pomyślnie"
            },
            "400": {
              "description": "Błędne dane"
            }
          }
        }
      }
    }
  }